BACKLOG-12495 : put the hook into an intermediate component in order to execute the query only when needed
